backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
anthropic>=0.23.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
moviepy>=1.0.3
yt-dlp>=2024.1.0
python-multipart>=0.0.6
aiofiles>=23.2.0
pillow>=10.0.0
numpy>=1.24.0
httpx>=0.26.0
soundfile>=0.12.1
# ElevenLabs TTS (API-based, no GPU required)
elevenlabs>=1.0.0
